refactor!: standardize on okio.ByteString — one byte vocabulary across the API
Wire-generated proto types force okio.ByteString into every proto-typed
call for both this SDK and its consumers (Meshtastic-Android consumes
the same protobufs artifact), so the kotlinx-io ByteString on Frame /
SessionPasskey was a gratuitous second vocabulary. Standardize:

- Frame.bytes and SessionPasskey.bytes are okio.ByteString.
- send(portnum, payload: kotlinx.io.Buffer) is replaced by
  send(portnum, payload: okio.ByteString) — the type proto payloads
  already arrive in, so payloads pass through without array copies.
- Internal frame assembly (WireCodec.FrameDecoder, TcpTransport,
  SerialFrameAssembler) moves to okio.Buffer.
- kotlinx-io is dropped from every published module (and the version
  catalog); okio is promoted to api in :core since it now appears in
  core's own public signatures.
- The toKotlinxByteString()/toOkioByteString() adapters added earlier
  this cycle are deleted — nothing left to bridge.

Consumers (and the app cutover) now use exactly one byte type
end-to-end: proto payloads, frames, and storage. ABI dumps regenerated;
docs/api-reference.md API-conventions section updated.

+++ core/src/commonMain/kotlin/org/meshtastic/sdk/WireCodec.kt
@@ -7,8 +7,7 @@
  */
 package org.meshtastic.sdk
 
-import kotlinx.io.Buffer
-import kotlinx.io.readByteArray
+import okio.Buffer
 import org.meshtastic.proto.FromRadio
 import org.meshtastic.proto.ToRadio
 
@@ -167,7 +166,7 @@ public object WireCodec {
                 }
 
                 State.READ_PAYLOAD -> {
-                    buffer.writeByte(b)
+                    buffer.writeByte(b.toInt())
                     bytesRemaining--
 
                     if (bytesRemaining == 0) {
